# swagger.yaml
openapi: 3.0.0
info:
  title: Психологическая Служба API
  description: API для сервиса психологической помощи
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Локальный сервер разработки
  - url: http://185.128.105.126:8000
    description: Сервер разработки
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Описание ошибки
    UserBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        middle_name:
          type: string
          nullable: true
        last_name:
          type: string
        phone_number:
          type: string
        email:
          type: string
          format: email
        social_media:
          type: string
          nullable: true
        password:
          type: string
          writeOnly: true
    UserCreateRequest:
      type: object
      required:
        - first_name
        - last_name
        - phone_number
        - email
        - password
        - role
      properties:
        first_name:
          type: string
          example: "Иван"
        middle_name:
          type: string
          nullable: true
          example: "Иванович"
        last_name:
          type: string
          example: "Иванов"
        phone_number:
          type: string
          example: "+79001234567"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        password:
          type: string
          minLength: 8
          example: "strongPassword123"
        role:
          type: string
          enum: [Student, Therapist, Administrator, Stuff]
          example: "Student"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "ivan@example.com"
        password:
          type: string
          example: "strongPassword123"
    TokenResponse:
      type: object
      properties:
        status_code:
          type: integer
          example: 200
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
paths:
  /users/user:
    get:
      tags:
        - users
      summary: Получить информацию о пользователе по email, id или токену.
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
            format: email
        - name: id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о пользователе. Если нет email или id, то вернется информация о текущем пользователе по токену из куки.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBase'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/register:
    post:
      tags:
        - users
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/login:
    post:
      tags:
        - users
      summary: Авторизация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...